chg:user:remove printing of list of images with nid and type at sync this code is specific of FIWARE and it is obsoleted when a different synchronisation algorithm is selected. This list has no sense, because it may not match with the list of images to synchronise, that anyway it depends of the target.
As consequence, the GlanceSync ordinary class is used now. The other
class was only used to invoke the print images method.

The funcionality of printing the images classified with nid and type
may be useful with FIWARE, but anyway it is available yet in the
examples folder as an independent utility.